Remove the PrimitiveProjection namespace enum in favor of -Projection suffixes
